requirements.txt
django==3.2
djangorestframework==3.12.4
gunicorn==20.1.0

# postgresql
psycopg2-binary==2.9.6

# elasticsearch
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22

# django-redis
redis==3.5.3
django-redis==5.0.0

# celery
celery[redis]>=5.0.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
flower==1.2.0